+++ src/picker/file_loader.py
@@ -1,3 +1,4 @@
+from glob import glob
 from typing import Dict, List
 
 from .paths import *
@@ -10,7 +11,7 @@ def read_characters_from_files(files: List[str], frecent: List[str], use_additio
         for character in frecent:
             all_characters[character] = []
 
-    for file in __resolve_all_files(files, use_additional):
+    for file in __resolve_all_filenames(files, use_additional):
         characters_from_file = __load_from_file(file)
         for line in characters_from_file:
             parsed_line = line.split(" ", 1)
@@ -23,36 +24,50 @@ def read_characters_from_files(files: List[str], frecent: List[str], use_additio
     return [f"{key} {value}" for key, value in collected_characters.items() if value != ""]
 
 
-def __resolve_all_files(files: List[str], use_additional: bool) -> List[Path]:
+def __resolve_all_filenames(file_names: List[str], use_additional: bool) -> List[Path]:
     resolved_file_names = []
-    for file_name in files:
-        resolved_file_names += __resolve_file(file_name, use_additional)
+    for file_name in file_names:
+        resolved_file_names += __resolve_filename(file_name, use_additional)
 
     return resolved_file_names
 
 
-def __resolve_file(file_name: str, use_additional: bool) -> List[Path]:
+def __resolve_filename(file_name: str, use_additional: bool) -> List[Path]:
     resolved_file_names = []
 
-    provided_file = Path(__file__).parent / "data" / f"{file_name}.csv"
+    for file in glob(file_name):
+        resolved_file_names.append(Path(file))
 
-    if Path(file_name).expanduser().is_file():
-        resolved_file_names.append(Path(file_name).expanduser())
-    elif provided_file.is_file():
-        resolved_file_names.append(provided_file)
-        custom_additional_file = custom_additional_files_location / f"{file_name}.additional.csv"
-        if custom_additional_file.is_file():
-            resolved_file_names.append(custom_additional_file)
-        provided_additional_file = Path(__file__).parent / "data" / "additional" / f"{file_name}.csv"
-        if use_additional and provided_additional_file.is_file():
-            resolved_file_names.append(provided_additional_file)
-    elif file_name == "all":
-        nested_file_names = [__resolve_file(file.stem) for file in (Path(__file__).parent / "data").glob("*.csv")]
+    if resolved_file_names:
+        return resolved_file_names
+
+    for file in (Path(__file__).parent / "data").glob(file_name if "*" in file_name else file_name + "*"):
+        resolved_file_names.append(file)
+        resolved_file_names += __load_addtional_files(file, use_additional)
+
+    if resolved_file_names:
+        return resolved_file_names
+
+    if file_name == "all":
+        nested_file_names = [
+            __resolve_filename(file.stem, use_additional) for file in (Path(__file__).parent / "data").glob("*.csv")
+        ]
         resolved_file_names += [file_name for file_names in nested_file_names for file_name in file_names]
-    else:
-        raise FileNotFoundError(f"Couldn't find file {file_name!r}")
+        return resolved_file_names
 
-    return resolved_file_names
+    raise FileNotFoundError(f"Couldn't find file {file_name!r}")
+
+
+def __load_addtional_files(original_file: Path, use_additional: bool) -> List[Path]:
+    additional_files = []
+    custom_additional_file = custom_additional_files_location / f"{original_file.stem}.additional.csv"
+    if custom_additional_file.is_file():
+        additional_files.append(custom_additional_file)
+    provided_additional_file = Path(__file__).parent / "data" / "additional" / f"{original_file.stem}.csv"
+    if use_additional and provided_additional_file.is_file():
+        additional_files.append(provided_additional_file)
+
+    return additional_files
 
 
 def __load_from_file(file: Path) -> List[str]: